MEGA_STRUCTURE_1000X/neilchat/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
websockets==12.0
redis==5.0.1
httpx==0.27.0
python-multipart==0.0.9
pydantic-settings==2.4.0
psutil==5.9.8
chromadb==0.4.24
pymupdf==1.23.26
beautifulsoup4==4.12.3
litellm>=1.20.0
PyJWT==2.8.0
prometheus-fastapi-instrumentator==7.0.0
numpy<2.0.0
